requirements.txt
asgiref==3.8.1
Django==4.2.7
django-userforeignkey==0.5.0
djangorestframework==3.14.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pytz==2024.1
sqlparse==0.4.4
